Refactor mock$JQueryElement
